# doc/API_Documentation.openapi.yaml
openapi: 3.1.0
...
info:
  title: Core User Management API
  version: 1.0.0-draft
  summary: Reusable OpenAPI 3.1 contract for account lifecycle, authentication, user profile management, MFA, passkeys, admin user operations, and privacy preferences.
  description: |
    Canonical OpenAPI 3.1 contract for the reusable core user-management layer in this repository.

    This spec intentionally excludes the PriceTracker business-domain APIs such as catalog, shops, prices,
    purchases, comparisons, watchlists, alerts, and moderation of price submissions.

    The retained surface is limited to reusable platform concerns:

    - registration, login, logout, refresh, password reset, and MFA completion
    - current-user profile, avatar attachment, and contact management
    - passkeys, TOTP MFA, recovery codes, and security summary
    - account-owned private file handling needed by the core, such as avatars or export packages
    - admin user management and system settings
    - legal document discovery, privacy consent capture, and cookie preferences

    Contract conventions:

    - IDs use `format: uuid`
    - timestamps use `format: date-time` in UTC
    - default currency uses ISO 4217 alpha-3 codes
    - response bodies use `camelCase`
    - successful responses use `{ data, meta }`
    - error responses use `{ error }` and include `urgencyLevel` on a 1-9 scale
    - request DTOs reject unknown fields unless explicitly stated otherwise
servers:
...
    description: Production
...
paths:
...
  /me/privacy-consents:
    get:
      tags:
        - Privacy
      operationId: listPrivacyConsents
      summary: List the caller accepted legal document versions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Accepted legal document versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacyConsentListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Privacy
      operationId: createPrivacyConsents
      summary: Record acceptance of updated legal documents
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacyConsentCreateRequest'
      responses:
        '201':
          description: Accepted document versions recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacyConsentListEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/privacy-requests:
    get:
      tags:
        - Privacy
      operationId: listPrivacyRequests
      summary: List privacy or account lifecycle requests submitted by the current account
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Privacy request list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacyRequestListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Privacy
      operationId: createPrivacyRequest
      summary: Create a privacy or account lifecycle request
      description: Submit requests such as access export, erasure, rectification, or account deletion.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrivacyRequestCreateRequest'
      responses:
        '201':
          description: Privacy request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacyRequestEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/privacy-requests/{privacyRequestId}:
    get:
      tags:
        - Privacy
      operationId: getPrivacyRequest
      summary: Get one privacy or account lifecycle request
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PrivacyRequestIdPath'
      responses:
        '200':
          description: Privacy request detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivacyRequestEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /privacy/cookie-preferences:
    get:
      tags:
        - Privacy
      operationId: getCookiePreferences
      summary: Get current cookie preference state
      description: Return the current cookie preference snapshot for the current browser or signed-in user context.
      responses:
        '200':
          description: Cookie preference snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CookiePreferencesEnvelope'
    put:
      tags:
        - Privacy
      operationId: setCookiePreferences
      summary: Set cookie preferences
      description: Update cookie preference flags. Necessary cookies are always enabled and are not configurable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CookiePreferencesUpdateRequest'
      responses:
        '200':
          description: Updated cookie preference snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CookiePreferencesEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 256
      example: cur_eyJvZmZzZXQiOjIwfQ
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
    PrivacyRequestIdPath:
      name: privacyRequestId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Identifier'
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: One or more fields are invalid.
                  urgencyLevel: 3
                  details:
                    field: email
                  requestId: req_01HZZZZZZZZZZ
    Unauthorized:
      description: Authentication is required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: UNAUTHORIZED
                  message: Authentication is required.
                  urgencyLevel: 4
                  requestId: req_01HZZZZZZZZZZ
...
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error:
                  code: NOT_FOUND
                  message: Resource not found.
                  urgencyLevel: 2
                  requestId: req_01HZZZZZZZZZZ
...
  schemas:
    Identifier:
      type: string
      format: uuid
      example: 11111111-1111-4111-8111-111111111111
...
    ResponseMeta:
      type: object
      additionalProperties: false
      required:
        - requestId
      properties:
        requestId:
          type: string
          example: req_01HZZZZZZZZZZ
        nextCursor:
          type:
            - string
            - 'null'
          example: null
    ErrorObject:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - urgencyLevel
        - requestId
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: One or more fields are invalid.
        urgencyLevel:
          type: integer
          minimum: 1
          maximum: 9
          description: |
            Integer urgency for the error response.
            Lower values are usually client-fixable request issues.
            Higher values indicate stronger security controls or backend/operator attention.
          example: 3
        details:
          type:
            - object
            - 'null'
          additionalProperties: true
        requestId:
          type: string
          example: req_01HZZZZZZZZZZ
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorObject'
...
    LegalDocumentAcceptance:
      type: object
      additionalProperties: false
      required:
        - documentKey
        - version
      properties:
        documentKey:
          type: string
          enum:
            - terms_of_service
            - privacy_policy
            - cookie_policy
          example: privacy_policy
        version:
          type: string
          example: '2026-04-01'
...
    PrivacyConsent:
      type: object
      additionalProperties: false
      required:
        - documentKey
        - version
        - acceptedAt
      properties:
        documentKey:
          type: string
          enum:
            - terms_of_service
            - privacy_policy
            - cookie_policy
          example: privacy_policy
        version:
          type: string
          example: '2026-04-01'
        acceptedAt:
          type: string
          format: date-time
          example: '2026-04-02T16:00:00Z'
        source:
          type:
            - string
            - 'null'
          enum:
            - registration
            - settings_update
            - admin_import
            - null
          example: registration
    PrivacyConsentCreateRequest:
      type: object
      additionalProperties: false
      required:
        - documents
      properties:
        documents:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LegalDocumentAcceptance'
    PrivacyConsentListEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PrivacyConsent'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    CookiePreferences:
      type: object
      additionalProperties: false
      required:
        - necessary
        - preferences
        - analytics
        - marketing
        - updatedAt
      properties:
        necessary:
          type: boolean
          const: true
          example: true
        preferences:
          type: boolean
          example: true
        analytics:
          type: boolean
          example: false
        marketing:
          type: boolean
          example: false
        updatedAt:
          type: string
          format: date-time
          example: '2026-04-02T16:10:00Z'
    CookiePreferencesUpdateRequest:
      type: object
      additionalProperties: false
      required:
        - preferences
        - analytics
        - marketing
      properties:
        preferences:
          type: boolean
          example: true
        analytics:
          type: boolean
          example: false
        marketing:
          type: boolean
          example: false
    CookiePreferencesEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/CookiePreferences'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    PrivacyRequest:
      type: object
      additionalProperties: false
      required:
        - id
        - requestType
        - status
        - requestedAt
      properties:
        id:
          $ref: '#/components/schemas/Identifier'
        requestType:
          type: string
          enum:
            - access_export
            - erasure
            - rectification
            - account_delete
          example: access_export
        status:
          type: string
          enum:
            - open
            - in_review
            - completed
            - rejected
            - cancelled
          example: open
        requestedAt:
          type: string
          format: date-time
          example: '2026-04-23T08:00:00Z'
        dueAt:
          type:
            - string
            - 'null'
          format: date-time
          example: '2026-05-23T08:00:00Z'
        completedAt:
          type:
            - string
            - 'null'
          format: date-time
          example: null
        notes:
          type:
            - string
            - 'null'
          example: Please include login history and verified contact data.
        exportFileId:
          type:
            - string
            - 'null'
          format: uuid
          example: null
    PrivacyRequestEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/PrivacyRequest'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    PrivacyRequestListEnvelope:
      type: object
      additionalProperties: false
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PrivacyRequest'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    PrivacyRequestCreateRequest:
      type: object
      additionalProperties: false
      required:
        - requestType
      properties:
        requestType:
          type: string
          enum:
            - access_export
            - erasure
            - rectification
            - account_delete
          example: account_delete
        notes:
          type:
            - string
            - 'null'
          maxLength: 1000
          example: Please delete my account after exporting my profile and session history.
...